PSARC 2007/591 Generic x86 Machine Check Architecture PSARC 2007/594 Intel CPU and 5000/7300 Series Chipset FMA 6443855 x86 modinfo slots being wasted during cpu module path search 6475380 cmi_load() should unload unused CPU module 6520280 MCA status bits not properly being cleared for intel processors 6558878 topo_node_hash returns out of range hash 6567218 FMA for Intel processors and 5000/7300 series memory controller 6567634 generic x86 MCA support needs to be improved 6607616 prepare FMA/x86 for xVM 6607626 eversholt Makefile.com lacks a default target 6607637 Add number of cpus and cores per chip to cpu_info kstat 6607643 x86 on_trap handler jumps from frying pan into the fire 6616180 topo should translate slashes in authority fields
--HG--
rename : usr/src/uts/i86pc/cpu/amd_opteron/ao_mc.c => deleted_files/usr/src/uts/i86pc/cpu/amd_opteron/ao_mc.c
rename : usr/src/uts/i86xpv/generic_cpu/Makefile => deleted_files/usr/src/uts/i86xpv/generic_cpu/Makefile

+++ usr/src/common/mc/mc-amd/mcamd_rowcol.c
@@ -18,7 +18,7 @@
  *
  * CDDL HEADER END
  *
- * Copyright 2006 Sun Microsystems, Inc.  All rights reserved.
+ * Copyright 2007 Sun Microsystems, Inc.  All rights reserved.
  * Use is subject to license terms.
  */
 
@@ -94,7 +94,7 @@ getcsprops(struct mcamd_hdl *hdl, mcamd_node_t *cs, const char *caller,
 		mcamd_dprintf(hdl, MCAMD_DBG_ERR, "%s: failed to read cs "
 		    "props for cs 0x%p\n", caller, cs);
 		return (mcamd_set_errno(hdl, EMCAMD_TREEINVALID));
-	    }
+	}
 
 	return (0);
 }
@@ -452,7 +452,7 @@ mc_pa_to_offset(struct mcamd_hdl *hdl, mcamd_node_t *mc, mcamd_node_t *cs,
 
 	offset_un.do_valid = 1;
 	offset_un.do_version = MCAMD_OFFSET_VERSION;
-	offset_un.do_rank = csp.dimmrank;
+	offset_un.do_rank = (uint32_t)csp.dimmrank;
 	offset_un.do_row = rowaddr;
 	offset_un.do_bank = bankaddr;
 	offset_un.do_col = coladdr;
